More realistic physics of motion.
Momentum is now preserved, so turning does not instantaneously change the velocity vector as it did before. Sufficient thrust must now also be applied.

This makes helm quite a bit more challenging.

Angular momentum is still not modeled, so turning is still trivial. Something to consider.

Includes a new showMotion debug option that draws the velocity vector of moving entities on the map.