Science gets a map.
Map can be panned with drag gesture, and does not follow the ship like the other stations.

Selected object can be scanned as before, with data now shown as map overlay.

Selected object is also shown on the map with an animated disc effect.

Also reduced opacity of grid lines to be less dominating.